fix(ci): sync mobile locales and trim DashboardView for SwiftLint
- Regenerate iOS/Android locale resources from canonical mobile locale sources
- Extract DashboardHeroPanel to bring DashboardView under type_body_length limit
