+++ clients/android/app/src/main/res/values-fr/strings.xml
@@ -933,24 +933,4 @@
         <item quantity="one">Vous avez révisé %1$d élément</item>
         <item quantity="other">Vous avez révisé %1$d éléments</item>
     </plurals>
-    <string name="mobile_drawer_dashboard">Tableau de bord</string>
-    <string name="mobile_drawer_todos">À faire</string>
-    <string name="mobile_drawer_review">Révision</string>
-    <string name="mobile_drawer_notebooks">Mes carnets</string>
-    <string name="mobile_drawer_globalNotebook">Carnet global</string>
-    <string name="mobile_drawer_accommodations">Aménagements</string>
-    <string name="mobile_drawer_back">Retour</string>
-    <string name="mobile_drawer_menu">Menu</string>
-    <string name="mobile_drawer_courseMenu">Menu du cours</string>
-    <string name="mobile_drawer_group_learning">Apprentissage</string>
-    <string name="mobile_drawer_group_notes">Notes et portfolio</string>
-    <string name="mobile_drawer_group_administration">Administration</string>
-    <string name="mobile_drawer_group_account">Compte</string>
-    <string name="mobile_drawer_group_teaching">Enseignement</string>
-    <string name="mobile_drawer_course_content">Contenu</string>
-    <string name="mobile_drawer_course_collaboration">Collaboration</string>
-    <string name="mobile_drawer_course_grades">Notes et analyses</string>
-    <string name="mobile_drawer_course_people">Personnes</string>
-    <string name="mobile_drawer_course_manage">Gérer</string>
-    <string name="mobile_drawer_pinned">Épinglés</string>
 </resources>
